Implement `archive-get` subcommand
Added a new `archive-get` subcommand which downloads an archive from
a given URL and optionally decompresses it with `zstd`. The download
content is streamed into the `zstd` decoder for efficiency. Users can
also choose the `zstd` window log size parameter used by the decoder,
equivalent to specifying `--long=$VALUE` for the `unzstd` util.

Signed-off-by: George Pisaltu <georgep@casperlabs.io>