Remove `checksum` option in `archive create`
The `archive create` subcommand had an option to disable the
requirement for enabling checksums on zstd frames in case the option
couldn't be set because of the environment. We have since decided to
scrap this option as it would not only be confusing for users but
also unnecessary since we have no reason to believe this fails in
any common use case and environment.

Signed-off-by: George Pisaltu <georgep@casperlabs.io>