Refactoring to use `thiserror` in codebase
This commit removes the old subcommand success reporting system where
subcommands would log errors and return a `bool`. Now all
subcommand errors are defined with `thiserror` and are all unified in
a big error enum at the `subcommand` mod level. This means errors
are properly propagated all the way to the top level and can be handled
in all lower levels as appropriate.

Signed-off-by: George Pisaltu <georgep@casperlabs.io>